adicionando scaffolding controller, migration, tarefa.cs

+++ Controllers/TarefasController.cs
@@ -1,7 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
 using aplicacao_asp.net_mvc.Data;
 using aplicacao_asp.net_mvc.Models;
-using Microsoft.EntityFrameworkCore;
 
 namespace aplicacao_asp.net_mvc.Controllers
 {
@@ -17,20 +22,25 @@ namespace aplicacao_asp.net_mvc.Controllers
         // GET: Tarefas
         public async Task<IActionResult> Index()
         {
-            var tarefas = await _context.Tarefas.Include(t => t.Tipo).ToListAsync();
-            return View(tarefas);
+            var applicationDbContext = _context.Tarefas.Include(t => t.Tipo);
+            return View(await applicationDbContext.ToListAsync());
         }
 
         // GET: Tarefas/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null) return NotFound();
+            if (id == null)
+            {
+                return NotFound();
+            }
 
             var tarefa = await _context.Tarefas
                 .Include(t => t.Tipo)
-                .FirstOrDefaultAsync(t => t.Id == id);
-
-            if (tarefa == null) return NotFound();
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (tarefa == null)
+            {
+                return NotFound();
+            }
 
             return View(tarefa);
         }
@@ -38,65 +48,95 @@ namespace aplicacao_asp.net_mvc.Controllers
         // GET: Tarefas/Create
         public IActionResult Create()
         {
-            ViewData["Tipos"] = _context.Tipos.ToList();
+            ViewData["TipoId"] = new SelectList(_context.Tipos, "Id", "Descricao");
             return View();
         }
 
         // POST: Tarefas/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(Tarefa tarefa)
+        public async Task<IActionResult> Create([Bind("Id,Nome,Descricao,DataCriacao,DataConclusao,TipoId")] Tarefa tarefa)
         {
             if (ModelState.IsValid)
             {
-                tarefa.DataCriacao = DateTime.Now;
                 _context.Add(tarefa);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["Tipos"] = _context.Tipos.ToList();
+            ViewData["TipoId"] = new SelectList(_context.Tipos, "Id", "Descricao", tarefa.TipoId);
             return View(tarefa);
         }
 
         // GET: Tarefas/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null) return NotFound();
+            if (id == null)
+            {
+                return NotFound();
+            }
 
             var tarefa = await _context.Tarefas.FindAsync(id);
-            if (tarefa == null) return NotFound();
-
-            ViewData["Tipos"] = _context.Tipos.ToList();
+            if (tarefa == null)
+            {
+                return NotFound();
+            }
+            ViewData["TipoId"] = new SelectList(_context.Tipos, "Id", "Descricao", tarefa.TipoId);
             return View(tarefa);
         }
 
         // POST: Tarefas/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, Tarefa tarefa)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,Descricao,DataCriacao,DataConclusao,TipoId")] Tarefa tarefa)
         {
-            if (id != tarefa.Id) return NotFound();
+            if (id != tarefa.Id)
+            {
+                return NotFound();
+            }
 
             if (ModelState.IsValid)
             {
-                _context.Update(tarefa);
-                await _context.SaveChangesAsync();
+                try
+                {
+                    _context.Update(tarefa);
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!TarefaExists(tarefa.Id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["Tipos"] = _context.Tipos.ToList();
+            ViewData["TipoId"] = new SelectList(_context.Tipos, "Id", "Descricao", tarefa.TipoId);
             return View(tarefa);
         }
 
         // GET: Tarefas/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null) return NotFound();
+            if (id == null)
+            {
+                return NotFound();
+            }
 
             var tarefa = await _context.Tarefas
                 .Include(t => t.Tipo)
-                .FirstOrDefaultAsync(t => t.Id == id);
-
-            if (tarefa == null) return NotFound();
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (tarefa == null)
+            {
+                return NotFound();
+            }
 
             return View(tarefa);
         }
@@ -107,9 +147,18 @@ namespace aplicacao_asp.net_mvc.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var tarefa = await _context.Tarefas.FindAsync(id);
-            _context.Tarefas.Remove(tarefa);
+            if (tarefa != null)
+            {
+                _context.Tarefas.Remove(tarefa);
+            }
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
+        }
+
+        private bool TarefaExists(int id)
+        {
+            return _context.Tarefas.Any(e => e.Id == id);
         }
     }
 }
